refactor: Code readability improvements
